+++ tests/smoke/playwright/test/upload-test-e2e.spec.ts
@@ -4,7 +4,7 @@ import {
   UploadTarget,
   getFileSelection,
   getUploadTargets
-} from '../resources/test-helpers';
+} from '../resources/test-utils';
 
 test.describe.configure({ mode: 'parallel' });
