+++ tests/smoke/playwright/test/upload-test-pages.spec.ts
@@ -5,7 +5,7 @@ import {
   UploadTarget,
   getFileSelection,
   getUploadTargets
-} from '../resources/test-helpers';
+} from '../resources/test-utils';
 
 test.describe.configure({ mode: 'parallel' });
