+++ upload-server/configs/testing/e2e/azurite_destinations.yml
