+++ upload-server/configs/testing/e2e/minio_destinations.yml
@@ -9,6 +9,7 @@ targets:
     secret_access_key: minioadmin
     endpoint: 'http://minio:8000'
     region: us-east-1
+    path_template: $EDAV_PATH_TEMPLATE
   ehdi:
     name: ehdi
     type: s3
@@ -17,6 +18,7 @@ targets:
     secret_access_key: minioadmin
     endpoint: 'http://minio:8000'
     region: us-east-1
+    path_template: $EHDI_PATH_TEMPLATE
   eicr:
     name: eicr
     type: s3
@@ -25,6 +27,7 @@ targets:
     secret_access_key: minioadmin
     endpoint: 'http://minio:8000'
     region: us-east-1
+    path_template: $EICR_PATH_TEMPLATE
   ncird:
     name: ncird
     type: s3
@@ -33,6 +36,7 @@ targets:
     secret_access_key: minioadmin
     endpoint: 'http://minio:8000'
     region: us-east-1
+    path_template: $NCIRD_PATH_TEMPLATE
 routing_groups:
   - data_stream_id: pulsenet
     data_stream_route: localsequencefile
@@ -42,13 +46,11 @@ routing_groups:
     data_stream_route: csv
     delivery_targets:
       - name: ehdi
-        path_template: $EHDI_PATH_TEMPLATE
   - data_stream_id: eicr
     data_stream_route: fhir
     delivery_targets:
       - name: eicr
-        path_template: $EICR_PATH_TEMPLATE
-  - data_stream_id: rsp-prevention
+  - data_stream_id: covid-all-monthly-vaccination
     data_stream_route: csv
     delivery_targets:
       - name: ncird
@@ -64,7 +66,7 @@ routing_groups:
     data_stream_route: other
     delivery_targets:
       - name: ncird
-  - data_stream_id: rsp-prevention
+  - data_stream_id: rsv-prevention
     data_stream_route: csv
     delivery_targets:
       - name: ncird
