+++ upload-server/docker-compose.e2e.all.yml
@@ -26,7 +26,7 @@ services:
       - UI_SERVER_EXTERNAL_HOST=upload-server-all:8080
       - UI_SERVER_INTERNAL_HOST=upload-server-all:8080
       - UPLOAD_CONFIG_PATH=/upload-configs
-      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/mix_destinations.yml
+      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/e2e/mix_destinations.yml
       - AZURITE_KEY=${AZURITE_STORAGE_KEY}
       - EDAV_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
       - EHDI_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
