+++ upload-server/docker-compose.e2e.azurite.yml
@@ -22,7 +22,7 @@ services:
       - AZURE_STORAGE_KEY=${AZURITE_STORAGE_KEY}
       - AZURE_ENDPOINT=http://azurite:10000/devstoreaccount1
       - TUS_AZURE_CONTAINER_NAME=test
-      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/azurite_destinations.yml
+      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/e2e/azurite_destinations.yml
       - AZURITE_KEY=${AZURITE_STORAGE_KEY}
       - EDAV_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
       - EHDI_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
