+++ upload-server/docker-compose.e2e.minio-azurite.yml
@@ -31,7 +31,7 @@ services:
       - AWS_ACCESS_KEY_ID=minioadmin
       - AWS_SECRET_ACCESS_KEY=minioadmin
       - DEX_MANIFEST_CONFIG_FOLDER_NAME=upload-configs
-      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/azurite_destinations.yml
+      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/e2e/azurite_destinations.yml
       - AZURITE_KEY=${AZURITE_STORAGE_KEY}
       - EDAV_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
       - EHDI_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
