+++ upload-server/docker-compose.e2e.yml
@@ -13,7 +13,7 @@ services:
       - UI_SERVER_EXTERNAL_HOST=upload-server-fs:8080
       - UI_SERVER_INTERNAL_HOST=upload-server-fs:8080
       - UPLOAD_CONFIG_PATH=/upload-configs
-      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/file_destinations.yml
+      - DEX_DELIVERY_CONFIG_FILE=/code/configs/testing/e2e/file_destinations.yml
       - EDAV_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
       - EHDI_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
       - EICR_PATH_TEMPLATE='{{.Filename}}_{{.UploadId}}'
